requirements.txt
streamlit
groq
pypdf
python-docx
scikit-learn
pandas
streamlit-authenticator